implement logic to show tagged users in bottom sheets
