+++ lib/views/screens/chats/messages/widgets/message_screen_widget.dart
@@ -2,162 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../../blocs/bloc.dart';
-import '../../../../../model/chat/chat_model.dart';
 import '../../../../../model/chat/message_model.dart';
 import '../../../../../model/pure_user_model.dart';
-import '../../../../../services/chat/chat_service.dart';
-import '../../../../../utils/navigate.dart';
-import '../../../../../utils/palette.dart';
 import '../../../../widgets/avatar.dart';
-import '../../../settings/profile/profile_screen.dart';
-import '../../group/group_info_screen.dart';
 import 'message_inbox_widget.dart';
 import 'messages_body.dart';
-
-class MessageAppBarTitle extends StatelessWidget {
-  final String chatId;
-  final PureUser receipient;
-  final bool hasPresenceActivated;
-  const MessageAppBarTitle({
-    Key? key,
-    required this.chatId,
-    required this.receipient,
-    this.hasPresenceActivated = false,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () => viewFullProfile(context, receipient),
-      child: Row(
-        children: [
-          Avartar(
-            size: 22,
-            ringSize: 0.8,
-            imageURL: receipient.photoURL,
-          ),
-          const SizedBox(width: 10.0),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  receipient.fullName,
-                  maxLines: 1,
-                  style: const TextStyle(
-                    fontWeight: FontWeight.w600,
-                    fontSize: 17.5,
-                    fontFamily: Palette.sanFontFamily,
-                    letterSpacing: 0.5,
-                  ),
-                ),
-                if (hasPresenceActivated)
-                  BlocBuilder<UserPresenceCubit, UserPresenceState>(
-                    builder: (context, state) {
-                      final status = state is UserPresenceSuccess &&
-                          state.presence.isOnline;
-                      if (status)
-                        return Padding(
-                          padding: const EdgeInsets.only(top: 2.0),
-                          child: Text(
-                            "Online",
-                            maxLines: 1,
-                            style: TextStyle(
-                              color: Theme.of(context)
-                                  .colorScheme
-                                  .primaryVariant
-                                  .withOpacity(0.6),
-                              fontWeight: FontWeight.w400,
-                              fontSize: 13.0,
-                              fontFamily: Palette.sanFontFamily,
-                              letterSpacing: 0.25,
-                            ),
-                          ),
-                        );
-                      return Offstage();
-                    },
-                  ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void viewFullProfile(BuildContext context, PureUser user) {
-    push(
-      context: context,
-      page: ProfileScreen(user: user, hideConnectionStatus: true),
-    );
-  }
-}
-
-class GroupMessageAppBarTitle extends StatefulWidget {
-  final ChatModel chat;
-  const GroupMessageAppBarTitle({Key? key, required this.chat})
-      : super(key: key);
-
-  @override
-  State<GroupMessageAppBarTitle> createState() =>
-      _GroupMessageAppBarTitleState();
-}
-
-class _GroupMessageAppBarTitleState extends State<GroupMessageAppBarTitle> {
-  late ChatModel chat;
-
-  @override
-  void initState() {
-    super.initState();
-    chat = widget.chat;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () => viewGroupProfile(context),
-      child: Row(
-        children: [
-          Avartar(size: 22, ringSize: 0.8, imageURL: chat.groupImage!),
-          const SizedBox(width: 10.0),
-          Text(
-            chat.groupName!,
-            maxLines: 1,
-            style: const TextStyle(
-              fontWeight: FontWeight.w600,
-              fontSize: 17.5,
-              fontFamily: Palette.sanFontFamily,
-              letterSpacing: 0.5,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> viewGroupProfile(BuildContext context) async {
-    final state = BlocProvider.of<GroupCubit>(context).state;
-    if (state is GroupMembers) {
-      List<PureUser> members = state.members;
-
-      Navigator.of(context).push<void>(
-        MaterialPageRoute(builder: (context) {
-          return MultiBlocProvider(
-            providers: [
-              BlocProvider(create: (_) => GroupChatCubit(ChatServiceImp())),
-              BlocProvider(create: (_) => ParticipantCubit(ChatServiceImp())),
-            ],
-            child: GroupInfoScreen(
-              chat: chat,
-              participants: members,
-              onChatChanged: (newChat) => setState(() => chat = newChat),
-            ),
-          );
-        }),
-      );
-    }
-  }
-}
 
 class MessageBody extends StatefulWidget {
   final String chatId;
@@ -171,10 +20,12 @@ class MessageBody extends StatefulWidget {
 
 class _MessageBodyState extends State<MessageBody> {
   final _inputFocusNode = FocusNode();
+  final _userTaggedNotifier = ValueNotifier<String?>(null);
 
   @override
   void dispose() {
     _inputFocusNode.dispose();
+    _userTaggedNotifier.dispose();
     super.dispose();
   }
 
@@ -190,12 +41,38 @@ class _MessageBodyState extends State<MessageBody> {
         children: [
           // Messages
           Expanded(
-            child: Messagesbody(
-              chatId: widget.chatId,
-              inputFocusNode: _inputFocusNode,
-              firstName: widget.receipientName,
-              onSentButtonPressed: (final message) =>
-                  sendMessage(context, message),
+            child: Stack(
+              fit: StackFit.expand,
+              children: [
+                Messagesbody(
+                  chatId: widget.chatId,
+                  inputFocusNode: _inputFocusNode,
+                  firstName: widget.receipientName,
+                  onSentButtonPressed: (final message) =>
+                      sendMessage(context, message),
+                ),
+                if (widget.receipientName == null)
+                  Align(
+                    alignment: Alignment.bottomCenter,
+                    child: BlocBuilder<GroupCubit, GroupState>(
+                      builder: (context, state) {
+                        if (state is GroupMembers) {
+                          return ValueListenableBuilder<String?>(
+                            valueListenable: _userTaggedNotifier,
+                            builder: (context, value, _) {
+                              if (value == null) return Offstage();
+                              final users = taggedUsers(state.members, value);
+                              return users.isEmpty
+                                  ? Offstage()
+                                  : _TaggedUsers(members: users);
+                            },
+                          );
+                        }
+                        return Offstage();
+                      },
+                    ),
+                  ),
+              ],
             ),
           ),
           // Message Input Box
@@ -205,7 +82,8 @@ class _MessageBodyState extends State<MessageBody> {
             // recipient name is null if it is a group chat
             // because the current user is not conversing with a specific user
             // but rather a group of users
-            allowUserTagging: widget.receipientName == null,
+            userTaggingNotifier:
+                widget.receipientName == null ? _userTaggedNotifier : null,
             onSentButtonPressed: (final message) =>
                 sendMessage(context, message),
           )
@@ -216,5 +94,40 @@ class _MessageBodyState extends State<MessageBody> {
 
   void sendMessage(final BuildContext context, final MessageModel message) {
     context.read<MessageCubit>().sendMessage(widget.chatId, message);
+  }
+
+  List<PureUser> taggedUsers(List<PureUser> users, String value) {
+    return users.toList().where((member) {
+      return member.fullName.toLowerCase().contains(value.toLowerCase());
+    }).toList();
+  }
+}
+
+class _TaggedUsers extends StatelessWidget {
+  final List<PureUser> members;
+  const _TaggedUsers({Key? key, required this.members}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DraggableScrollableSheet(
+      minChildSize: 0.4,
+      builder: (context, scrollController) {
+        return Container(
+          color: Theme.of(context).dialogBackgroundColor,
+          child: ListView.separated(
+            itemCount: members.length,
+            controller: scrollController,
+            separatorBuilder: (_, __) => Divider(height: 0),
+            itemBuilder: (context, index) {
+              final member = members[index];
+              return ListTile(
+                leading: Avartar2(imageURL: member.photoURL),
+                title: Text(member.fullName),
+              );
+            },
+          ),
+        );
+      },
+    );
   }
 }
